Cast MaxHealth to float for TestEqual

+++ Source/autobattlerTests/Private/Test_CombatSystem.cpp
@@ -29,7 +29,7 @@ bool FCombatSystemTest::RunTest(const FString& Parameters)
     TestUnit->DispatchBeginPlay();
 
     FUnitData* TestWarriorData = UnitDataTable->FindRow<FUnitData>(FName("TestWarrior"), "");
-    TestEqual(TEXT("Unit's health should be initialized from DataTable"), TestUnit->CurrentHealth, TestWarriorData->MaxHealth);
+    TestEqual(TEXT("Unit's health should be initialized from DataTable"), TestUnit->CurrentHealth, static_cast<float>(TestWarriorData->MaxHealth));
 
     return true;
 }
